Update README.md
Summary
This portfolio is a well-rounded and interactive showcase of a front-end developer's skills and services. By combining modern web design techniques with dynamic and engaging features, it provides a professional and polished online presence. The use of HTML, CSS, and JavaScript ensures a robust and user-friendly experience, making it an effective tool for demonstrating the developer's capabilities and expertise.